+++ src/Actions/ProductSuperAttributeActionInterface.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * TechDivision\Import\Product\Variant\Repositories\ProductSuperLinkAction
+ * TechDivision\Import\Product\Variant\Repositories\ProductSuperAttributeActionInterface
  *
  * NOTICE OF LICENSE
  *
@@ -20,10 +20,10 @@
 
 namespace TechDivision\Import\Product\Variant\Actions;
 
-use TechDivision\Import\Actions\AbstractAction;
+use TechDivision\Import\Actions\ActionInterface;
 
 /**
- * An action implementation that provides CRUD functionality for product super links.
+ * Inerface for action implementations that provides CRUD functionality for product super attributes.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
  * @copyright 2016 TechDivision GmbH <info@techdivision.com>
@@ -31,6 +31,6 @@ use TechDivision\Import\Actions\AbstractAction;
  * @link      https://github.com/techdivision/import-product-variant
  * @link      http://www.techdivision.com
  */
-class ProductSuperLinkAction extends AbstractAction implements ProductSuperLinkActionInterface
+interface ProductSuperAttributeActionInterface extends ActionInterface
 {
 }