+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.0.0-beta8
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Add missing interfaces for actions and repositories
+* Replace class type hints for ProductVariantProcessor with interfaces
+
 # Version 1.0.0-beta7
 
 ## Bugfixes